# remove prefetches from cryptonight scratchpad inner loop