@fireice-uk 2017-01-12
That is one interesting result - I wrote this code 2 months ago, but I distinctly remember that prefetching into L1 was faster than just reminding the CPU to keep it in L3. 
There is a reason why caching is referred to as the "dark arts" :). And every architecture behaves differently.  

I'm fairly busy now but if you are able test results from Sandy Bridge and Westmere cores would be nice to have. My gut feeling is that we stumbled onto an updated prefetching algorithm in Haswell.