@kevinlekiller 2017-01-13
I tested this a few days back on cpuminer-opt (noticed it was commented out in the original cryptonote / bytecoin code : https://github.com/cryptonotefoundation/cryptonote/blob/master/src/crypto/slow-hash.inl#L96).

I get about 1-3 more H/s per core(6700k 3 threads), but when I use the computer, for browsing for example, my H/s dips further than when uncommented, it usually dips down consistently to about 230-240H/s when doing stuff, but with commented out it's very variable, it ranges from 150 to 230 (even with the highest CPU priority).

I also got about 2-3 more H/s per core on cpuminer-opt by profiling (-fprofile-generate). By experimenting, it seems to be -funroll-loops that made the most difference.