@jquesnelle 2017-01-13
I don't have access to another CPU right now but I'll try to investigate more. Just my analysis (could be totally wrong) but I don't believe that the `_mm_prefetch` command is actually instructing the CPU to _keep_ the address in L1, rather to simply move it there until it's evicted naturally (per [MSDN](https://msdn.microsoft.com/en-us/library/84szxsww(v=vs.100).aspx): Loads one cache line of data from address p to a location closer to the processor). In our case the address would be moved to L1 naturally anyway as a result of the access that occurs immediately after the prefetch:

```
_mm_prefetch(&l0[idx0 & 0x1FFFF0], _MM_HINT_T0);

uint64_t hi, lo, cl, ch;
cl = ((uint64_t*)&l0[idx0 & 0x1FFFF0])[0];
```

There doesn't appear to be any substantial work that's ever performed between the prefetch and the access, which would be my guess as to why it's more of a hindrance than a help. Although it's equally likely that this is some new Haswell (and above) specific prefetch algorithm we're bumping up against.