@fireice-uk 2017-01-13
I tested it a bit more, and it appears to be due to large pages, rather than cpu core - removing prefetches on slow memory murders performance (-20%).

I can't accept the patch as it is (as i get enough "why isn't is faster on slow memory"), but I will rework it to have a separate, no-prefetch, large page hashing core.