@dimecoin 2017-01-13
I tested this on 4 servers, my hash rate whent from average of 277.5 down to 271.2

All have same CPU and all have large page support.  Let me know if you need more info:

    Intel(R) Core(TM) i7-2600 CPU @ 3.40GHz