@fireice-uk 2017-01-14
On my systems I seem to get a modest 2-3% boost on Westmere cores, and a bigger 5-6% on Sandy and Haswell. I think I will add this to a per-thread config as one of the settings that you can play around with to optimize the miner to your build.

Dimecoin, does the server have buffered memory by any chance?